+++ src/avp/types/call_serial_number.rs
@@ -8,10 +8,10 @@ pub struct CallSerialNumber {
 
 impl CallSerialNumber {
     const ATTRIBUTE_TYPE: u16 = 15;
-    const LENGTH: u16 = 4;
+    const LENGTH: usize = 4;
 
     pub fn try_read<'a, 'b>(reader: &'b mut impl Reader<'a>) -> ResultStr<Self> {
-        if reader.len() < Self::LENGTH as usize {
+        if reader.len() < Self::LENGTH {
             return Err("Incomplete CallSerialNumber AVP encountered");
         }
 
@@ -27,7 +27,7 @@ impl From<u32> for CallSerialNumber {
 }
 
 impl QueryableAVP for CallSerialNumber {
-    fn get_length(&self) -> u16 {
+    fn get_length(&self) -> usize {
         Self::LENGTH
     }
 }
