+++ src/avp/types/called_number.rs
@@ -23,8 +23,8 @@ impl CalledNumber {
 }
 
 impl QueryableAVP for CalledNumber {
-    fn get_length(&self) -> u16 {
-        self.value.len() as u16
+    fn get_length(&self) -> usize {
+        self.value.len()
     }
 }
