+++ src/avp/types/calling_number.rs
@@ -23,8 +23,8 @@ impl CallingNumber {
 }
 
 impl QueryableAVP for CallingNumber {
-    fn get_length(&self) -> u16 {
-        self.value.len() as u16
+    fn get_length(&self) -> usize {
+        self.value.len()
     }
 }
