+++ src/avp/types/proxy_authen_response.rs
@@ -19,8 +19,8 @@ impl ProxyAuthenResponse {
 }
 
 impl QueryableAVP for ProxyAuthenResponse {
-    fn get_length(&self) -> u16 {
-        self.data.len() as u16
+    fn get_length(&self) -> usize {
+        self.data.len()
     }
 }
